updated index.html
added some styling to the font of the h1 and p elements, as well as updated their margins. I also set the attribute field to the bottom and updated the div tag to footer for better accessibility. I also added border radius to the image and main container per the styling, but have to figure out how to give the image the look of it being outlined by the main container.