updated index.html
Fixed and adjusted the main container so that it would border the qr code image. I also adjusted the border-radius of the image to be 50% of the border-radius of the main container and adjusted the padding around it. I also adjusted the padding and margin around the H1 and P element to match the design visually. Lastly, I added a drop shadow behind the main container, the original design looked to have a slight one.